Removed the old implementation.
